AI-Enhanced-Document-QA-System-Script/requirements.txt
python-dotenv
pinecone-client
bertopic
transformers
torch
